Playlist

+++ src/main/java/nl/han/oose/dea/spotitube/presentation/PlaylistService.java
@@ -3,6 +3,7 @@ package nl.han.oose.dea.spotitube.presentation;
 import nl.han.oose.dea.spotitube.data.AbstractMapper;
 import nl.han.oose.dea.spotitube.data.PlaylistDataMapper;
 import nl.han.oose.dea.spotitube.domain.Playlist;
+import nl.han.oose.dea.spotitube.domain.User;
 
 import javax.inject.Inject;
 import javax.ws.rs.*;
@@ -13,6 +14,7 @@ import java.sql.ResultSet;
 public class PlaylistService {
 
     private AbstractMapper dataMapper;
+    private static final int MOCK_DURATION = 1234;
 
     @Inject
     public void setDataMapper(PlaylistDataMapper dataMapper) {
@@ -22,9 +24,10 @@ public class PlaylistService {
     @GET
     @Produces("application/json")
     public Response getAllPlaylists(@QueryParam("token") String token) {
+        if(token == null) return Response.status(400).build();
         return Response
                 .status(200)
-                .entity(new PlaylistsDTO(dataMapper.findAll(), 1234))
+                .entity(new PlaylistsDTO(dataMapper.findAll(), MOCK_DURATION))
                 .build();
     }
 
@@ -32,10 +35,24 @@ public class PlaylistService {
     @DELETE
     @Produces("application/json")
     public Response deletePlaylist(@QueryParam("token") String token, @PathParam("id") int id) {
+        if(token == null) return Response.status(400).build();
         dataMapper.delete(id);
         return Response
                 .status(200)
-                .entity(dataMapper.findAll())
+                .entity(new PlaylistsDTO(dataMapper.findAll(), MOCK_DURATION))
                 .build();
+    }
+
+    @POST
+    @Consumes("application/json")
+    @Produces("application/json")
+    public Response addPlaylist(@QueryParam("token") String token, Playlist playlist) {
+        if(token == null) return Response.status(400).build();
+        dataMapper.add(playlist);
+        return Response
+                .status(201)
+                .entity(new PlaylistsDTO(dataMapper.findAll(), MOCK_DURATION))
+                .build();
+
     }
 }
